requirements.txt
Flask==1.1.1
mysql==0.0.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
requests==2.22.0
gunicorn==20.0.4
Flask-Cors==3.0.8
beautifulsoup4==4.8.0